# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MediTrack Pro API — PULSO IoT Pill Dispensing System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients/{patientId}/medications:
    get:
      operationId: listMedications
      tags: [medications]
      summary: List medications for a patient
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of medications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medication"
    post:
      operationId: createMedication
      tags: [medications]
      summary: Add a medication
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
        "409":
          description: Drug interaction warning
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InteractionWarning"
  /patients/{patientId}/medications/{medicationId}:
    get:
      operationId: getMedication
      tags: [medications]
      summary: Get a medication
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
        - name: medicationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Medication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMedication
      tags: [medications]
      summary: Update a medication
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
        - name: medicationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
    delete:
      operationId: deleteMedication
      tags: [medications]
      summary: Delete a medication
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
        - name: medicationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Dosing Schedules ───────────────────────────────────────
  /patients/{patientId}/medications/{medicationId}/schedules:
    get:
      operationId: listSchedules
      tags: [schedules]
      summary: List dosing schedules for a medication
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
        - name: medicationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Schedule"
    post:
      operationId: createSchedule
      tags: [schedules]
      summary: Create a dosing schedule
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
        - name: medicationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
  /patients/{patientId}/medications/{medicationId}/schedules/{scheduleId}:
    patch:
      operationId: updateSchedule
      tags: [schedules]
      summary: Update a dosing schedule
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
        - name: medicationId
          in: path
          required: true
          schema:
            type: integer
        - name: scheduleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
    delete:
      operationId: deleteSchedule
      tags: [schedules]
      summary: Delete a dosing schedule
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
        - name: medicationId
          in: path
          required: true
          schema:
            type: integer
        - name: scheduleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Devices ────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # Auth
...
    MedicationInput:
      type: object
      required: [name, dosage, funnelNumber, stockCount]
      properties:
        name:
          type: string
        genericName:
          type: string
        dosage:
          type: string
        form:
          type: string
          enum: [tablet, capsule, liquid, other]
          default: tablet
        funnelNumber:
          type: integer
          minimum: 1
          maximum: 8
        stockCount:
          type: integer
        instructions:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
    MedicationUpdate:
      type: object
      properties:
        name:
          type: string
        genericName:
          type: string
        dosage:
          type: string
        form:
          type: string
          enum: [tablet, capsule, liquid, other]
        funnelNumber:
          type: integer
        stockCount:
          type: integer
        instructions:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
    Medication:
      type: object
      required: [id, patientId, name, dosage, funnelNumber, stockCount, daysRemaining]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        name:
          type: string
        genericName:
          type: ["string", "null"]
        dosage:
          type: string
        form:
          type: string
        funnelNumber:
          type: integer
        stockCount:
          type: integer
        stockPercentage:
          type: number
        daysRemaining:
          type: ["integer", "null"]
        instructions:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        isLowStock:
          type: boolean
        createdAt:
          type: string
          format: date-time
    InteractionWarning:
      type: object
      required: [warning, conflictingMedications]
      properties:
        warning:
          type: string
        conflictingMedications:
          type: array
          items:
            type: string
        canOverride:
          type: boolean

    # Schedule
    ScheduleInput:
      type: object
      required: [timeType]
      properties:
        timeType:
          type: string
          enum: [clock, meal_relative]
        clockTime:
          type: ["string", "null"]
        mealRelation:
          type: ["string", "null"]
          enum: [before_breakfast, after_breakfast, before_lunch, after_lunch, before_dinner, after_dinner, "null"]
        daysOfWeek:
          type: array
          items:
            type: string
            enum: [mon, tue, wed, thu, fri, sat, sun]
    ScheduleUpdate:
      type: object
      properties:
        timeType:
          type: string
          enum: [clock, meal_relative]
        clockTime:
          type: ["string", "null"]
        mealRelation:
          type: ["string", "null"]
        daysOfWeek:
          type: array
          items:
            type: string
    Schedule:
      type: object
      required: [id, medicationId, timeType]
      properties:
        id:
          type: integer
        medicationId:
          type: integer
        timeType:
          type: string
          enum: [clock, meal_relative]
        clockTime:
          type: ["string", "null"]
        mealRelation:
          type: ["string", "null"]
        daysOfWeek:
          type: array
          items:
            type: string
        nextDoseAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time

    # Device
...